+++ src/diagnostics/html/validators/index.ts
@@ -4,6 +4,7 @@ export { DivValidator } from './div';
 export { FieldsetValidator } from './fieldset';
 export { HeadingValidator } from './heading';
 export { InputValidator } from './input';
+export { ImageValidator } from './image';
 export { LinkValidator } from './link';
 export { NavigationValidator } from './navigation';
 export { RequiredValidator } from './required';